ultra speed =)

+++ upload/system/library/bus_cache/mem.php
@@ -1,9 +1,7 @@
 <?php
 // *   Аўтар: "БуслікДрэў" ( https://buslikdrev.by/ )
-// *   © 2016-2022; BuslikDrev - Усе правы захаваныя.
-// *   Спецыяльна для сайта: "OpenCart.pro" ( https://opencart.pro/ )
+// *   © 2016-2024; BuslikDrev - Усе правы захаваны.
 
-//https://ospanel.io/forum/viewtopic.php?f=3&t=2191&start=10
 namespace Bus_Cache;
 //namespace Opencart\Extension\Bus_Cache\System\library\Bus_Cache;
 
@@ -13,11 +11,16 @@ if (!defined('VERSION')) {
 	exit('ЗАПРЫШЧАЮ!');
 }
 
+//https://ospanel.io/forum/viewtopic.php?f=3&t=2191&start=10
 class Mem {
 	private $expire;
+	private $active = false;
 	private $memcache;
 
 	public function __construct($expire = 3600) {
+		if (!defined('CACHE_PREFIX')) {
+			define('CACHE_PREFIX', 'cache_');
+		}
 		if (!defined('CACHEDUMP_LIMIT')) {
 			define('CACHEDUMP_LIMIT', 9999);
 		}
@@ -27,37 +30,59 @@ class Mem {
 		if (!defined('CACHE_PORT')) {
 			define('CACHE_PORT', 11211);
 		}
-		if (!defined('CACHE_PREFIX')) {
-			define('CACHE_PREFIX', 'cache_');
+		if (!defined('CACHE_NULL')) {
+			define('CACHE_NULL', false);
 		}
+
 		$this->expire = $expire;
-		$this->memcache = new \Memcache();
-		$this->memcache->pconnect(CACHE_HOSTNAME, CACHE_PORT);
+
+		$this->active = extension_loaded('memcache') && class_exists('Memcache') && function_exists('memcache_connect');
+
+		if ($this->active) {
+			$this->memcache = new \Memcache();
+			$this->memcache->pconnect(CACHE_HOSTNAME, CACHE_PORT);
+		}
 	}
 
 	public function get($key) {
-		return $this->memcache->get(CACHE_PREFIX . $key);
+		if ($this->active) {
+			return $this->memcache->get(CACHE_PREFIX . $key);
+		} else {
+			return CACHE_NULL;
+		}
 	}
 
-	public function set($key, $value, $expire = 3600) {
-		return $this->memcache->set(CACHE_PREFIX . $key, $value, MEMCACHE_COMPRESSED, $this->expire);
+	public function set($key, $value, $expire = 0) {
+		if ($this->active) {
+			if (!$expire) {
+				$expire = $this->expire;
+			}
+
+			$this->memcache->set(CACHE_PREFIX . $key, $value, MEMCACHE_COMPRESSED, $expire);
+		}
 	}
 
 	public function delete($key) {
-		$this->memcache->delete(CACHE_PREFIX . $key);
+		if ($this->active) {
+			if ($key == '*') {
+				$this->flush();
+			} else {
+				$this->memcache->delete(CACHE_PREFIX . $key);
+			}
+		}
 	}
 
 	public function flush($timer = 5) {
-		$status = false;
-
-		if (method_exists($this->memcache, 'flush')) {
-			$this->memcache->flush();
-			$status = true;
+		if ($this->active) {
+			return $this->memcache->flush();
+		} else {
+			return false;
 		}
-		if (method_exists($this->memcache, 'close')) {
+	}
+
+	public function __destruct() {
+		if ($this->active) {
 			$this->memcache->close();
 		}
-
-		return $status;
 	}
 }
