+++ upload/system/library/bus_cache/memcached.php
@@ -1,9 +1,7 @@
 <?php
 // *   Аўтар: "БуслікДрэў" ( https://buslikdrev.by/ )
-// *   © 2016-2022; BuslikDrev - Усе правы захаваныя.
-// *   Спецыяльна для сайта: "OpenCart.pro" ( https://opencart.pro/ )
+// *   © 2016-2024; BuslikDrev - Усе правы захаваны.
 
-//https://ospanel.io/forum/viewtopic.php?f=3&t=2191&start=10
 namespace Bus_Cache;
 //namespace Opencart\Extension\Bus_Cache\System\library\Bus_Cache;
 
@@ -13,11 +11,22 @@ if (!defined('VERSION')) {
 	exit('ЗАПРЫШЧАЮ!');
 }
 
+//https://ospanel.io/forum/viewtopic.php?f=3&t=2191&start=10
 class Memcached {
 	private $expire;
+	private $active = false;
 	private $memcached;
 
 	public function __construct($expire = 3600) {
+		if (!defined('CACHE_PREFIX')) {
+			define('CACHE_PREFIX', 'cache_');
+		}
+		if (!defined('CACHE_USERNAME')) {
+			define('CACHE_USERNAME', '');
+		}
+		if (!defined('CACHE_PASSWORD')) {
+			define('CACHE_PASSWORD', '');
+		}
 		if (!defined('CACHEDUMP_LIMIT')) {
 			define('CACHEDUMP_LIMIT', 9999);
 		}
@@ -27,38 +36,60 @@ class Memcached {
 		if (!defined('CACHE_PORT')) {
 			define('CACHE_PORT', 11211);
 		}
-		if (!defined('CACHE_PREFIX')) {
-			define('CACHE_PREFIX', 'cache_');
+		if (!defined('CACHE_NULL')) {
+			define('CACHE_NULL', false);
 		}
+
 		$this->expire = $expire;
-		$this->memcached = new \Memcached();
-		$this->memcached->addServer(CACHE_HOSTNAME, CACHE_PORT);
-		//$this->memcached->setSaslAuthData(CACHE_USERNAME, CACHE_PASSWORD) ;
+
+		$this->active = extension_loaded('memcached') && class_exists('Memcached');
+
+		if ($this->active) {
+			$this->memcached = new \Memcached();
+			$this->memcached->addServer(CACHE_HOSTNAME, CACHE_PORT);
+			//$this->memcached->setSaslAuthData(CACHE_USERNAME, CACHE_PASSWORD) ;
+		}
 	}
 
 	public function get($key) {
-		return $this->memcached->get(CACHE_PREFIX . $key);
+		if ($this->active) {
+			return $this->memcached->get(CACHE_PREFIX . $key);
+		} else {
+			return CACHE_NULL;
+		}
 	}
 
 	public function set($key, $value, $expire = 0) {
-		return $this->memcached->set(CACHE_PREFIX . $key, $value, $this->expire);
+		if ($this->active) {
+			if (!$expire) {
+				$expire = $this->expire;
+			}
+
+			$this->memcached->set(CACHE_PREFIX . $key, $value, $expire);
+		}
 	}
 
 	public function delete($key) {
-		$this->memcached->delete(CACHE_PREFIX . $key);
+		if ($this->active) {
+			if ($key == '*') {
+				$this->flush(1);
+			} else {
+				$this->memcached->delete(CACHE_PREFIX . $key);
+			}
+		}
 	}
 
 	public function flush($timer = 5) {
-		$status = false;
-
-		if (method_exists($this->memcached, 'flush')) {
-			$this->memcached->flush($timer);
-			$status = true;
+		if ($this->active) {
+			return $this->memcached->flush($timer);
+		} else {
+			return false;
 		}
-		if (method_exists($this->memcached, 'quit')) {
+	}
+
+	public function __destruct() {
+		if ($this->active && method_exists($this->memcached, 'quit')) {
 			$this->memcached->quit();
 		}
-
-		return $status;
 	}
 }
